リファクタリング UIの共通部分の統一 ファイルとフォルダの整理
