+++ fueoni_ver2/lib/components/room/error_handling.dart
