+++ fueoni_ver2/lib/components/room_waiting/waiting.dart
@@ -1,20 +1,5 @@
 import 'package:flutter/material.dart';
 
-Widget roomIdDisplay({
-  required BuildContext context,
-  required int? roomId,
-}) {
-  return Padding(
-    padding: const EdgeInsets.all(16.0),
-    child: Text(
-      'ルームID: ${roomId ?? '生成中...'}',
-      style: const TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-      textAlign: TextAlign.center,
-    ),
-  );
-}
-
-// ユーザーリストを表示するウィジェット
 class UserList extends StatelessWidget {
   static const List<String> users = ['ユーザー1', 'ユーザー2', 'ユーザー3'];
