+++ fueoni_ver2/lib/screens/room_creation/room_creation_waiting_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:fueoni_ver2/components/room/room.dart';
 import 'package:fueoni_ver2/components/room_waiting/waiting.dart';
 import 'package:fueoni_ver2/services/creation_room_services.dart';
 
@@ -36,7 +37,7 @@ class RoomCreationWaitingScreen extends StatelessWidget {
       body: Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: <Widget>[
-          roomIdDisplay(context: context, roomId: args.roomId),
+          RoomWidgets.displayRoomId(roomId: args.roomId),
           const UserList(),
         ],
       ),
