--- fueoni_ver2/lib/screens/room_creation/widgets/passcode_display.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PasscodeDisplay extends StatelessWidget {
-  final String passcode;
-
-  const PasscodeDisplay({super.key, required this.passcode});
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(passcode.isNotEmpty ? '設定済み: $passcode' : 'パスコードを設定してください');
-  }
-}
