+++ fueoni_ver2/lib/screens/room_creation/widgets/timer_dialog.dart
@@ -15,6 +15,21 @@ Future<Duration?> showTimerDialog({
   );
 }
 
+class TimerDisplay extends StatelessWidget {
+  final Duration duration;
+
+  const TimerDisplay({super.key, required this.duration});
+
+  @override
+  Widget build(BuildContext context) {
+    String hours = duration.inHours.toString().padLeft(2, '0');
+    String minutes = (duration.inMinutes % 60).toString().padLeft(2, '0');
+    String seconds = (duration.inSeconds % 60).toString().padLeft(2, '0');
+
+    return Text('$hours:$minutes:$seconds');
+  }
+}
+
 class _TimerDialog extends StatefulWidget {
   final Duration initialDuration;
